Integrating spherical wrist detection into general 6R class
