+++ CPP/src/IK/CMakeLists.txt
@@ -9,7 +9,7 @@ endif()
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_FLAGS_RELEASE "-O3")
 
-set(sources Spherical_IK.cpp General_IK.cpp 3R_IK.cpp ${CMAKE_CURRENT_SOURCE_DIR}/utils/kinematic_utils.cpp ${CMAKE_CURRENT_SOURCE_DIR}/../utils/kinematic_remodelling.cpp)
+set(sources General_IK.cpp 3R_IK.cpp ${CMAKE_CURRENT_SOURCE_DIR}/utils/kinematic_utils.cpp ${CMAKE_CURRENT_SOURCE_DIR}/../utils/kinematic_remodelling.cpp)
 
 add_library(EAIK_Kinematics ${sources})
 target_include_directories(EAIK_Kinematics PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_SOURCE_DIR}/../utils ${CMAKE_CURRENT_SOURCE_DIR}/../../external/ik-geo/cpp/subproblems)
